requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
marshmallow==3.20.1
python-multipart==0.0.6
requests==2.31.0
openai==1.3.0
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2